fix: address review findings
Round-4 multi-agent /review fixes:

- local_mip: count incumbent / construction branches in worker()
  too, so warm-start counters cover the bandit dispatch (port/det,
  port/opp), not just seq/* (R1-6 / R2-3 / R3-2)
- local_mip: WarmStartCounters use int64_t (R3-14)
- local_mip.h: clarify counter semantics — `construction` covers
  both fresh sweeps and cold-start cache hits, and the bandit's
  caller-supplied initial_solution does not increment any counter
- scylla_worker: drop redundant `ncol_ == 0` check after the new
  abort guard renders it unreachable (R2-4)
- local_mip_worker: skip non-finite solution[j] in perturb_solution
  (NaN / inf would propagate through the int64 cast, R2-1)
- local_mip_worker: clarify that `irange < 1` guard is a defense
  for future refactors that drop integer rounding (R3-12)
- contested_pdlp: document that publish_snapshot_for_test requires
  the caller to hold mu_ (acquire_for_test pattern); locking
  internally would deadlock the test fixtures
- CLAUDE.md: note that LocalMIP's effort_per_ms now folds in
  cold-start construction; kWeightLocalMip calibration is mildly
  stale on instances where construction fires (R3-13)

+++ src/local_mip.h
@@ -15,7 +15,16 @@ namespace local_mip {
 // branches correspond to the three #75 cold-start fallback rungs:
 //   - `pool`   : pool.copy_best succeeded (warm-start from #74 pool).
 //   - `incumbent` : pool was empty, mipdata->incumbent picked up.
-//   - `construction` : both empty, paper's construction phase ran.
+//   - `construction` : pool and incumbent both empty; either the
+//                      paper's construction phase ran or its
+//                      cold-start cache was reused (a single first-
+//                      worker construction can amortise across N
+//                      peers via `cold_start_cache`).
+// Counters cover the `run_parallel_*` paths only; the bandit-mode
+// `worker()` entry point (used by `mip_heuristic_portfolio`) also
+// increments these as of R4 review, but restart-callback warm-starts
+// inside the parallel loops are NOT counted (their work happens after
+// the initial start has already been resolved).
 // Call `reset_warm_start_counters()` before a HiGHS run, then read
 // `warm_start_counters()` after to assert which path actually fired.
 // Used by R1-8 / R2-7 / R3-3 round-3 review tests to distinguish #74
@@ -23,9 +32,9 @@ namespace local_mip {
 // the integration tests can't tell those paths apart since both
 // produce non-zero effort.
 struct WarmStartCounters {
-    int pool;
-    int incumbent;
-    int construction;
+    int64_t pool;
+    int64_t incumbent;
+    int64_t construction;
 };
 
 void reset_warm_start_counters();
